+++ profiles/stac-server/main.tf
@@ -12,6 +12,8 @@ module "stac-server" {
   enable_transactions_extension               = var.stac_server_inputs.enable_transactions_extension
   enable_collections_authx                    = var.stac_server_inputs.enable_collections_authx
   enable_ingest_action_truncate               = var.stac_server_inputs.enable_ingest_action_truncate
+  enable_response_compression                 = var.stac_server_inputs.enable_response_compression
+  items_max_limit                             = var.stac_server_inputs.items_max_limit
   collection_to_index_mappings                = var.stac_server_inputs.collection_to_index_mappings
   opensearch_version                          = var.stac_server_inputs.opensearch_version
   opensearch_cluster_instance_type            = var.stac_server_inputs.opensearch_cluster_instance_type